Fix dtype mismatch in compute_logits and add embed_tokens LoRA config option
- Fix RuntimeError in LoRAEmbedding.compute_logits(): cast weights to
  match hidden dtype (float32 vs bfloat16)
- Fix same dtype issue in _compute_lm_logits() fallback path
- Add commented embed_tokens option to skip-list in both YAML configs,
  allowing users to disable LoRAEmbedding and use full-train instead

+++ src/openpi/waypoint/joint_model.py
@@ -139,7 +139,7 @@ class PI0WaypointJoint(nn.Module):
         embed_layer = self.paligemma_with_expert.paligemma.language_model.embed_tokens
         if hasattr(embed_layer, "compute_logits"):
             return embed_layer.compute_logits(hidden)
-        return F.linear(hidden, embed_layer.weight)
+        return F.linear(hidden, embed_layer.weight.to(hidden.dtype))
 
     # ------------------------------------------------------------------
     # DDP-compatible forward dispatcher
